Padronização dos retornos com IActionResult

+++ Controllers/HomeController.cs
@@ -8,36 +8,41 @@ namespace Todo.Controllers
   public class HomeController : ControllerBase
   {
     [HttpGet("/")]
-    public List<TodoList> Get([FromServices] AppDataContext context)
+    public IActionResult Get([FromServices] AppDataContext context)
     {
-      return context.Todos.ToList();
+      return Ok(context.Todos.ToList());
     }
 
 
     [HttpPost("/")]
-    public TodoList Post(
+    public IActionResult Post(
       [FromBody] TodoList todo,
       [FromServices] AppDataContext context
     )
     {
       context.Todos.Add(todo);
       context.SaveChanges();
-      return todo;
+      return Created($"/{todo.Id}", todo);
     }
 
 
     [HttpGet("/{id:int}")]
-    public TodoList GetById(
+    public IActionResult GetById(
       [FromRoute] int id,
       [FromServices] AppDataContext context
     )
     {
-      return context.Todos.FirstOrDefault(x => x.Id == id);
+      var todo = context.Todos.FirstOrDefault(x => x.Id == id);
+
+      if (todo == null)
+        return NotFound();
+
+      return Ok(todo);
     }
 
 
     [HttpPut("/{id:int}")]
-    public TodoList Put(
+    public IActionResult Put(
       [FromRoute] int id,
       [FromBody] TodoList request,
       [FromServices] AppDataContext context
@@ -46,28 +51,31 @@ namespace Todo.Controllers
       var model = context.Todos.FirstOrDefault(x => x.Id == id);
 
       if (model == null)
-        return request;
+        return NotFound();
 
       model.Title = request.Title;
       model.Done = request.Done;
 
       context.Todos.Update(model);
       context.SaveChanges();
-      return model;
+      return Ok(model);
     }
 
 
     [HttpDelete("/{id:int}")]
-    public TodoList Delete(
+    public IActionResult Delete(
       [FromRoute] int id,
       [FromServices] AppDataContext context
     )
     {
       var model = context.Todos.FirstOrDefault(x => x.Id == id);
 
+      if (model == null)
+        return NotFound();
+
       context.Todos.Remove(model);
       context.SaveChanges();
-      return model;
+      return Ok(model);
     }
 
   }
